get working transfers
